v1.3
added in more assets and work on the random object allocator. Created barrier around the play area. Removed noise to terrain, as it was difficult to place objects properly. Needed more advanced algorithms that I currently don't have time to learn for the assignment (marching cubes) and would require a rework of my mesh generator.