v1.7
Added a start, instructions, options, pause Menu. A game over screen when an enemy gets to close and a win screen when the player collects all the notes. Controller Inputs have been added in game, while the enemy now spawns after 30s to allow players to move about and get use to movement and gameplay.